Move prune_branches.py from cases/git to devs
It's dev-workflow tooling like b64d/b64e, not a personal-life script
like mapit/pinterest, so it fits the devs/cases split better there.
Updated the console-script entry, test import, and README section
(now under devs, alphabetically after B64d/B64e).
